requirements.txt
PyYAML>=3.12
Pillow>=4.3.0